add siwe

+++ pages/api/reviews/[id].ts
@@ -1,6 +1,29 @@
 import { NextApiRequest, NextApiResponse } from 'next'
+import { getToken } from 'next-auth/jwt'
 import { ApiResponse, Review } from '../../../types'
 import { supabaseClient } from '../../../utils/supabase'
+
+const secret = process.env.NEXTAUTH_SECRET
+
+async function getReviewById(
+  review_id: string
+): Promise<Review | null> {
+  try {
+    const { data, error } = await supabaseClient
+      .from('reviews')
+      .select('*')
+      .eq('id', review_id)
+
+    if (error) {
+      throw error
+    }
+
+    return data[0] as Review || null
+  } catch (error) {
+    console.error('Erreur lors de la récupération des avis:', error)
+    return null
+  }
+}
 
 async function updateReview(
   reviewId: string,
@@ -42,17 +65,25 @@ export default async function handler(
 ) {
   const { query, method } = req
   const { id: review_id } = query
-
+  const token = await getToken({ req, secret })
   if (method === 'PUT') {
     if (review_id && typeof review_id === 'string') {
       const updatedReviewData: Partial<Review> = req.body
-
+      if (token?.sub !== updatedReviewData.user_id) {
+        res.status(403).json({ data: [], error: 'Not authorized' })
+        return
+      }
       const updatedReview = await updateReview(review_id, updatedReviewData)
       res.status(updatedReview ? 200 : 400).json({ data: updatedReview })
       return
     }
   } else if (method === 'DELETE') {
     if (review_id && typeof review_id === 'string') {
+      const review = await getReviewById(review_id)
+      if (!review || token?.sub !== review.user_id) {
+        res.status(403).json({ data: [], error: 'Not authorized' })
+        return
+      }
       const success = await deleteReview(review_id)
       res.status(success ? 200 : 400).json({ data: success })
       return
